backend/requirements.txt
Django>=5.0
django-ninja>=1.0
django-ninja-jwt>=1.0
django-cors-headers>=4.3.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
dj-database-url>=2.1.0
Pillow>=10.0.0
email-validator>=2.0.0
requests>=2.31.0
markdown>=3.5.0
weasyprint>=60.0
# Note: Use Python 3.11 or 3.12 (not 3.14) due to django-ninja-jwt Pydantic v1 compatibility

